+++ notifications/src/sdk/models/operations/get_benefits_plan_enrollment.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import planenrollment as shared_planenrollment
+from ...models.shared import planenrollment as shared_planenrollment
 from typing import List, Optional
 
 
@@ -11,11 +11,11 @@ from typing import List, Optional
 class GetBenefitsPlanEnrollmentResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    plan_enrollments: Optional[List[shared_planenrollment.PlanEnrollment]] = dataclasses.field(default=None)
+    classes: Optional[List[shared_planenrollment.PlanEnrollment]] = dataclasses.field(default=None)
     r"""List of plan enrollments for a member"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""