+++ notifications/src/sdk/sdkconfiguration.py
@@ -2,15 +2,16 @@
 
 import requests
 from dataclasses import dataclass
-from typing import Dict, Tuple
+from typing import Dict, Tuple, Callable, Union
 from .utils.retries import RetryConfig
 from .utils import utils
+from sdk.models import shared
 
 
 SERVERS = [
-    'https://api.wingspan.app/benefits',
+    'https://api.wingspan.app',
     # Wingspan's Benefits API in Production
-    'https://stagingapi.wingspan.app/benefits',
+    'https://stagingapi.wingspan.app',
     # Wingspan's Benefits API in Staging
 ]
 """Contains the list of servers available to the SDK"""
@@ -18,14 +19,14 @@ SERVERS = [
 @dataclass
 class SDKConfiguration:
     client: requests.Session
-    security_client: requests.Session
+    security: Union[shared.Security,Callable[[], shared.Security]] = None
     server_url: str = ''
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.0.0'
-    gen_version: str = '2.173.0'
-    user_agent: str = 'speakeasy-sdk/python 1.0.0 2.173.0 1.0.0 wingspan_notifications'
+    sdk_version: str = '3.0.1'
+    gen_version: str = '2.237.3'
+    user_agent: str = 'speakeasy-sdk/python 3.0.1 2.237.3 1.0.0 wingspan_notifications'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: